+++ coloraide/everything.py
@@ -35,6 +35,7 @@ from .spaces.cam16_ucs import CAM16UCS, CAM16LCD, CAM16SCD
 from .spaces.hct import HCT
 from .spaces.ucs import UCS
 from .spaces.rec709 import Rec709
+from .spaces.ryb import RYB, RYBBiased
 from .distance.delta_e_itp import DEITP
 from .distance.delta_e_99o import DE99o
 from .distance.delta_e_z import DEZ
@@ -94,6 +95,8 @@ ColorAll.register(
         CAM16LCD(),
         HCT(),
         UCS(),
+        RYB(),
+        RYBBiased(),
 
         # Delta E
         DEITP(),