+++ docs/src/py/notebook.py
@@ -553,7 +553,7 @@ def colorize(src, lang, **options):
 def color_command_validator(language, inputs, options, attrs, md):
     """Color validator."""
 
-    valid_inputs = {'exceptions', 'play'}
+    valid_inputs = {'exceptions', 'play', 'wheel'}
 
     for k, v in inputs.items():
         if k in valid_inputs:
@@ -657,6 +657,7 @@ def _color_command_formatter(src="", language="", class_name=None, options=None,
 
     # Support the new way
     gamut = kwargs.get('gamut', WEBSPACE)
+    wheel = options.get('wheel', False)
     play = options.get('play', False) if options is not None else False
     # Support the old way
     if not play and language == 'playground':
@@ -673,26 +674,93 @@ def _color_command_formatter(src="", language="", class_name=None, options=None,
         )
 
     try:
-        if len(md.preprocessors['fenced_code_block'].extension.stash) == 0:
-            code_id = 0
+        if wheel:
+            exceptions = options.get('exceptions', False) if options is not None else False
 
-        # Check if we should allow exceptions
-        exceptions = options.get('exceptions', False) if options is not None else False
+            _, colors = execute(src.strip(), not exceptions, init=init)
 
-        console, colors = execute(src.strip(), not exceptions, init=init)
-        el = _color_command_console(colors, gamut=gamut)
+            l = len(colors)
+            if l not in (12, 24, 48):
+                raise SuperFencesException("Color wheel requires either 12, 24, or 48 colors")
 
-        el += md.preprocessors['fenced_code_block'].extension.superfences[0]['formatter'](
-            src=console,
-            class_name="highlight",
-            language='pycon',
-            md=md,
-            options=options,
-            **kwargs
-        )
-        el = '<div class="color-command">{}</div>'.format(el)
-        el = template.format(el_id=code_id, raw_source=_escape(src), results=el, gamut=gamut)
-        code_id += 1
+            colors = [c[0].color for c in colors]
+
+            if l == 12:
+                freq = 4
+                offset = 6
+            elif l == 24:
+                freq = 8
+                offset = 12
+            else:
+                freq = 16
+                offset = 24
+
+            primary = list(reversed(colors[::freq]))
+            secondary = list(reversed(colors[offset::freq] + [colors[offset // 3]]))
+            tertiary = list(reversed(colors[::offset // 6]))
+            color_rings = [primary, secondary, tertiary]
+
+            extra_rings_start = ''
+            extra_rings_end = ''
+            if l > 12:
+                extra_rings_start = '<div class="tertiary2">'
+                extra_rings_end += '</div>'
+                color_rings.append(list(reversed(colors[::offset // 12])))
+            if l > 24:
+                extra_rings_start = '<div class="tertiary3">' + extra_rings_start
+                extra_rings_end += '</div>'
+                color_rings.append(list(reversed(colors)))
+
+            color_stops = ''
+            for i, colors in enumerate(color_rings, 1):
+                total = len(colors)
+                percent = 100 / total
+                current = percent
+                last = -1
+
+                stops = []
+                for e, color in enumerate(colors):
+                    color.fit(gamut)
+                    color_str = color.convert(gamut).to_string()
+                    if current:
+                        stops.append('{} {}%'.format(color_str, str(last)))
+                        stops.append('{} {}%'.format(color_str, str(current)))
+                        last = current
+                        if e < (total - 1):
+                            current += percent
+                        else:
+                            current = 100
+                    else:
+                        stops.append(color_str)
+                color_stops += "--color-wheel-stops{}: {};".format(i, ','.join(stops))
+
+            color_wheel = """<div class="color-wheel" style="{}"><div class="wheel">
+{}<div class="tertiary"><div class="secondary"><div class="secondary-inner"><div class="primary"><div class="primary-inner"></div></div></div></div></div></div></div>{}""" .format(  # noqa: E501
+                color_stops, extra_rings_start, extra_rings_end
+            )
+            return color_wheel
+
+        else:
+            if len(md.preprocessors['fenced_code_block'].extension.stash) == 0:
+                code_id = 0
+
+            # Check if we should allow exceptions
+            exceptions = options.get('exceptions', False) if options is not None else False
+
+            console, colors = execute(src.strip(), not exceptions, init=init)
+            el = _color_command_console(colors, gamut=gamut)
+
+            el += md.preprocessors['fenced_code_block'].extension.superfences[0]['formatter'](
+                src=console,
+                class_name="highlight",
+                language='pycon',
+                md=md,
+                options=options,
+                **kwargs
+            )
+            el = '<div class="color-command">{}</div>'.format(el)
+            el = template.format(el_id=code_id, raw_source=_escape(src), results=el, gamut=gamut)
+            code_id += 1
     except SuperFencesException:
         raise
     except Exception: