+++ tests/test_roundtrip.py
@@ -19,6 +19,8 @@ class TestRoundTrip:
         """Local color object."""
 
     Color.deregister('space:hpluv')
+    Color.deregister('space:ryb')
+    Color.deregister('space:ryb-biased')
 
     SPACES = {k: 5 for k in Color.CS_MAP.keys()}
     # Not as accurate due to bisect back to CAM16
@@ -88,6 +90,8 @@ class TestAchromaticRoundTrip(TestRoundTrip):
         """Local color object."""
 
     Color.deregister('space:hpluv')
+    Color.deregister('space:ryb')
+    Color.deregister('space:ryb-biased')
 
     SPACES = {k: 5 for k in Color.CS_MAP.keys()}
     # Precision just isn't as high for these in achromatic region
@@ -99,6 +103,82 @@ class TestAchromaticRoundTrip(TestRoundTrip):
         Color('gray'),
         Color('black'),
         Color('white')
+    ]
+
+    @pytest.mark.parametrize('space', SPACES)
+    def test_round_trip(self, space):
+        """Test round trip."""
+
+        for c in self.COLORS:
+            self.assert_round_trip(c, space)
+
+
+class TestRYBRoundTrip(TestRoundTrip):
+    """
+    Test RYB round trip.
+
+    We are only interested in testing this against sRGB.
+    """
+
+    class Color(Base):
+        """Local color object."""
+
+    SPACES = {'srgb': 5}
+
+    COLORS = [
+        Color('ryb', [1, 1, 1]),
+        Color('ryb', [1, 0, 0]),
+        Color('ryb', [1, 0, 0]),
+        Color('ryb', [0, 1, 0]),
+        Color('ryb', [0, 0, 1]),
+        Color('ryb', [1, 1, 0]),
+        Color('ryb', [0, 1, 1]),
+        Color('ryb', [1, 0, 1]),
+        Color('ryb', [0, 0, 0]),
+        Color('ryb', [0.2, 5, 0.7]),
+        Color('ryb', [0.8, 0.65, 0.21]),
+        Color('ryb', [0.8, 0.65, 0.21]),
+        Color('ryb', [0.8, 0.65, 0.21]),
+        Color('ryb', [0.9, 0.8, 0.21]),
+        Color('ryb', [0.1, 0.3, 0.05])
+    ]
+
+    @pytest.mark.parametrize('space', SPACES)
+    def test_round_trip(self, space):
+        """Test round trip."""
+
+        for c in self.COLORS:
+            self.assert_round_trip(c, space)
+
+
+class TestRYBBiasedRoundTrip(TestRoundTrip):
+    """
+    Test RYB round trip.
+
+    We are only interested in testing this against sRGB.
+    """
+
+    class Color(Base):
+        """Local color object."""
+
+    SPACES = {'srgb': 5}
+
+    COLORS = [
+        Color('ryb-biased', [1, 1, 1]),
+        Color('ryb-biased', [1, 0, 0]),
+        Color('ryb-biased', [1, 0, 0]),
+        Color('ryb-biased', [0, 1, 0]),
+        Color('ryb-biased', [0, 0, 1]),
+        Color('ryb-biased', [1, 1, 0]),
+        Color('ryb-biased', [0, 1, 1]),
+        Color('ryb-biased', [1, 0, 1]),
+        Color('ryb-biased', [0, 0, 0]),
+        Color('ryb-biased', [0.2, 5, 0.7]),
+        Color('ryb-biased', [0.8, 0.65, 0.21]),
+        Color('ryb-biased', [0.8, 0.65, 0.21]),
+        Color('ryb-biased', [0.8, 0.65, 0.21]),
+        Color('ryb-biased', [0.9, 0.8, 0.21]),
+        Color('ryb-biased', [0.1, 0.3, 0.05])
     ]
 
     @pytest.mark.parametrize('space', SPACES)